+++ spec/integration/sequence_and_naming.rb
@@ -15,6 +15,7 @@ class Hash
 end
 
 DEPENDENCIES = {
+  ConnectionTemplate: [:EthernetNetwork],
   Datacenter: [:Rack],
   DriveEnclosure: [:Enclosure],
   Enclosure: [:EnclosureGroup],
@@ -136,6 +137,7 @@ UPLINK_SET_NAME = 'EthernetUplinkSet_1'.freeze
 UPLINK_SET2_NAME = 'FCUplinkSet_1'.freeze
 UPLINK_SET3_NAME = 'FCUplinkSet_2'.freeze
 UPLINK_SET4_NAME = 'EthernetUplinkSet_2'.freeze
+UPLINK_SET5_NAME = 'FCUplinkSet_3'.freeze
 
 # LIGUplinkSet
 LIG_UPLINK_SET_NAME = 'EthernetUplinkSet_1'.freeze
@@ -240,6 +242,7 @@ SAS_INTERCONNECT1_NAME = "#{ENCLOSURE_1}, interconnect 1".freeze
 
 # USER
 USER_NAME = 'TestUser'.freeze
+FULL_NAME = 'Test User'.freeze
 
 # OS Deployment Plan
 OS_DEPLOYMENT_PLAN_NAME = 'HPE - Developer 1.0 - Deployment Test (UEFI)'.freeze